requirements.txt
asgiref==3.6.0
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-phonenumber-field==7.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-rest-swagger==2.2.0
drf-yasg==1.21.7
fcm-django==1.0.14
firebase-admin==5.4.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
httplib2==0.22.0
idna==3.4
# ipython==8.14.0
msgpack==1.0.5
oauthlib==3.2.2
phonenumbers==8.12.54
proto-plus==1.22.2
protobuf==4.23.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2023.3
pycountry==22.3.5
requests==2.29.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
pandas==2.0.3
geonamescache==2.0.0
python-magic==0.4.27
geopy==2.3.0
